Refactor: 함수 및 구조 분리
